Added new class - HandleGuess
